+++ demo/app-module.ts
@@ -1,32 +1,31 @@
 import { NgModule } from "@angular/core";
-import { HttpModule } from "@angular/http";
+import { HttpClientModule } from "@angular/common/http";
 import { FormsModule } from "@angular/forms";
-import { MdInputModule, MdListModule } from "@angular/material";
+import { MatInputModule, MatListModule } from "@angular/material";
 import { BrowserModule } from "@angular/platform-browser";
-import {BrowserAnimationsModule} from "@angular/platform-browser/animations";
+import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
 import { CollapseModule } from "ng2-bootstrap/collapse";
 
 import { AppComponent } from "./app-cmp";
 import { MaterialCmp } from "./material-cmp";
 import { NativeCmp } from "./native-cmp";
 import { routing } from "./app.routing";
-import { Ng2CompleterModule } from "../src";
+import { Ng2CompleterModule } from "../src/ng2-completer";
 import { CompleterCmpMd } from "./completer-cmp-md";
 
-
 @NgModule({
+  bootstrap: [AppComponent],
+  declarations: [AppComponent, CompleterCmpMd, NativeCmp, MaterialCmp],
   imports: [
     BrowserModule,
     BrowserAnimationsModule,
     Ng2CompleterModule,
-    MdInputModule,
-    MdListModule,
+    MatInputModule,
+    MatListModule,
     FormsModule,
-    HttpModule,
+    HttpClientModule,
     routing,
     CollapseModule
-  ],
-  declarations: [AppComponent, CompleterCmpMd, NativeCmp, MaterialCmp],
-  bootstrap: [AppComponent]
+  ]
 })
 export class AppModule { }