+++ demo/material-cmp.ts
@@ -1,7 +1,7 @@
 "use strict";
 import { Component } from "@angular/core";
 
-import { CompleterService, CompleterData } from "../src";
+import { CompleterService, CompleterData } from "../src/ng2-completer";
 
 let template = require("./material-cmp.html");
 let style = require("./material-cmp.css");
@@ -14,7 +14,7 @@ let style = require("./material-cmp.css");
 export class MaterialCmp {
     public countries = require("./res/data/countries.json");
 
-    private dataService: CompleterData;
+    public dataService: CompleterData;
 
     constructor(completerService: CompleterService) {
         this.dataService = completerService.local(this.countries, "name", "name").imageField("flag");